Move sendonce type to rtstat pkg, add download and build sends
